refactor(header): move header-bar to own component
