+++ src/App.vue
@@ -3,17 +3,15 @@ import '../node_modules/ol/ol.css'
 
 import { watch } from 'vue'
 import { storeToRefs } from 'pinia'
-import { useTranslation } from 'i18next-vue'
 
+import HeaderBar from './components/header/header-bar.vue'
 import BackgroundSelector from '@/components/background-selector/background-selector.vue'
-import LanguageSelector from '@/components/nav-bars/language-selector.vue'
+
 import LayerPanel from '@/components/layer-panel/layer-panel.vue'
 import MapContainer from '@/components/map/map-container.vue'
-import { themeSelectorService } from '@/components/theme-selector/theme-selector.service'
 
 import { statePersistorLayersService } from '@/services/state-persistor/state-persistor-layers.service'
 import { statePersistorThemeService } from '@/services/state-persistor/state-persistor-theme.service'
-import { useThemeStore } from '@/stores/config.store'
 import { statePersistorLayersOpenService } from '@/services/state-persistor/state-persistor-layersopen.service'
 import { useAppStore } from '@/stores/app.store'
 import useMap from './composables/map/map.composable'
@@ -22,23 +20,7 @@ statePersistorLayersService.bootstrap()
 statePersistorThemeService.bootstrap()
 statePersistorLayersOpenService.bootstrapLayersOpen()
 
-const { t } = useTranslation()
-const themeStore = useThemeStore()
-const { theme } = storeToRefs(themeStore)
-
-watch(
-  theme,
-  theme => {
-    if (theme) {
-      themeSelectorService.setCurrentThemeColors(theme)
-    }
-  },
-  { immediate: true }
-)
-
-const appStore = useAppStore()
-const { layersOpen } = storeToRefs(appStore)
-const { setLayersOpen } = appStore
+const { layersOpen } = storeToRefs(useAppStore())
 
 watch(layersOpen, () =>
   setTimeout(() => {
@@ -49,36 +31,7 @@ watch(layersOpen, () =>
 
 <template>
   <div class="h-screen flex flex-col overflow-hidden">
-    <header
-      class="lux-navbar w-full h-14 flex bg-white shadow-header z-10 shrink-0"
-    >
-      <div class="flex-2 p-[5px]">
-        <img src="./assets/header/gov-light.8b5db4.png" />
-      </div>
-      <div class="grow text-center">search</div>
-      <div>
-        <ul class="h-full flex">
-          <li>
-            <button
-              class="flex items-center before:font-icons before:text-3xl before:w-16 text-primary uppercase h-full mr-3"
-              :class="`before:content-${theme?.name}`"
-              data-cy="selectedThemeIcon"
-              @click="() => setLayersOpen(!layersOpen)"
-            >
-              <span class="hidden lg:inline-block">{{
-                t(`${theme?.name}`)
-              }}</span>
-            </button>
-          </li>
-          <li class="border-l-[1px] border-stone-300 h-full">
-            <language-selector
-              data-cy="langSelect"
-              class="flex-none h-full"
-            ></language-selector>
-          </li>
-        </ul>
-      </div>
-    </header>
+    <header-bar></header-bar>
     <main class="flex grow">
       <!--side bar-->
       <div v-if="layersOpen" class="w-80 bg-secondary">
